+++ src/main/java/com/nexmo/client/voice/endpoints/StopTalkMethod.java
@@ -29,6 +29,7 @@ import org.apache.commons.logging.Log;
 import org.apache.commons.logging.LogFactory;
 import org.apache.http.HttpResponse;
 import org.apache.http.client.methods.RequestBuilder;
+import org.apache.http.impl.client.BasicResponseHandler;
 import org.apache.http.util.EntityUtils;
 
 import java.io.IOException;
@@ -59,8 +60,7 @@ public class StopTalkMethod extends AbstractMethod<String, TalkResponse> {
 
     @Override
     public TalkResponse parseResponse(HttpResponse response) throws IOException {
-        String json = EntityUtils.toString(response.getEntity());
-        LOG.info("Received: " + json);
+        String json = new BasicResponseHandler().handleResponse(response);
         return TalkResponse.fromJson(json);
     }