+++ src/test/java/com/nexmo/client/verify/VerifyClientTest.java
@@ -154,7 +154,7 @@ public class VerifyClientTest {
     @Test
     public void testCheckInvalidResponse() throws Exception {
         wrapper.setHttpClient(this.stubHttpClient(
-                500,
+                200,
                 "<?xml version='1.0' encoding='UTF-8' ?>\n" +
                         "<not_valid_response />"));
 
@@ -390,7 +390,7 @@ public class VerifyClientTest {
 
     @Test
     public void testSearchHttpError() throws Exception {
-        wrapper.setHttpClient(this.stubHttpClient(500, "<?xml version=\"1.0\" encoding=\"UTF-8\"?>\n" +
+        wrapper.setHttpClient(this.stubHttpClient(200, "<?xml version=\"1.0\" encoding=\"UTF-8\"?>\n" +
                 "<verify_request />"));
         try {
             client.search("a-random-request-id");